Initial version

+++ src/Profilers/MonitorProfiler/Communication/CommandServer.cpp
@@ -55,7 +55,7 @@ void CommandServer::ListeningThread()
 {
     // TODO: Handle oom scenarios
     IpcMessage response;
-    response.Command = IpcCommand::Status;
+    response.Command = static_cast<unsigned short>(IpcCommand::Status);
     response.Payload.resize(sizeof(HRESULT));
 
     while (true)
